requirements.txt
django~=3.1.3
djangorestframework~=3.12.1
psycopg2-binary==2.8.6
pandas==1.1.3
python-dotenv==0.15.0
gunicorn==20.0.4
markdown==3.3.3
drf-yasg==1.20.0
freezegun~=1.0.0
